requirements.txt
langchain
langchain-community
langchain-google-genai
streamlit
pymupdf
pdfplumber 
pillow